+++ lib/active_record/connection_adapters/sqlserver_adapter.rb
@@ -481,15 +481,12 @@ module ActiveRecord
       end
 
       def version_year
-        @version_year ||= begin
-          if sqlserver_version =~ /vNext/
+        @version_year ||=
+          if /vNext/.match?(sqlserver_version)
             2016
           else
             /SQL Server (\d+)/.match(sqlserver_version).to_a.last.to_s.to_i
           end
-        rescue StandardError
-          2016
-        end
       end
 
       def sqlserver_version
@@ -499,7 +496,10 @@ module ActiveRecord
       private
 
       def connect
-        @raw_connection = self.class.new_client(@connection_parameters)
+        @sqlserver_version ||= begin
+          verify!
+          _raw_select("SELECT @@version", @raw_connection).first.first.to_s
+        end
       end
 
       def configure_connection