+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+#### Fixed
+
+- [#1366](https://github.com/rails-sqlserver/activerecord-sqlserver-adapter/pull/1366) Verify connection before retrieving the database version.
+
 ## v8.0.8
 
 #### Changed